Renamed Dom.js to CalendarDom.js
